# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YouTube Audience Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface VideoInput {
  /** YouTube video URL */
  url: string;
}

export type VideoStatus = typeof VideoStatus[keyof typeof VideoStatus];


export const VideoStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  error: 'error',
} as const;

export interface Video {
  id: number;
  youtubeId: string;
  title: string;
  channelName: string;
  viewCount: number;
  likeCount: number;
  publishedAt: string;
  thumbnailUrl: string;
  commentCount: number;
  status: VideoStatus;
  createdAt: string;
}

export type VideoDetailStatus = typeof VideoDetailStatus[keyof typeof VideoDetailStatus];


export const VideoDetailStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  error: 'error',
} as const;

export interface SentimentSummary {
  positive: number;
  negative: number;
  neutral: number;
  total: number;
  positivePercent?: number;
  negativePercent?: number;
  neutralPercent?: number;
}

export interface VideoDetail {
  id: number;
  youtubeId: string;
  title: string;
  channelName: string;
  viewCount: number;
  likeCount: number;
  publishedAt: string;
  thumbnailUrl: string;
  commentCount: number;
  status: VideoDetailStatus;
  createdAt: string;
  sentimentSummary: SentimentSummary;
}

export type CommentSentiment = typeof CommentSentiment[keyof typeof CommentSentiment];


export const CommentSentiment = {
  positive: 'positive',
  negative: 'negative',
  neutral: 'neutral',
} as const;

export interface Comment {
  id: number;
  videoId: number;
  text: string;
  authorName: string;
  likeCount: number;
  publishedAt: string;
  sentiment: CommentSentiment;
  confidence: number;
}

export interface CommentPage {
  data: Comment[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface SentimentTimePoint {
  date: string;
  positive: number;
  negative: number;
  neutral: number;
}

export interface Keyword {
  word: string;
  count: number;
}

export interface Analytics {
  sentimentSummary: SentimentSummary;
  sentimentOverTime: SentimentTimePoint[];
  topPositiveKeywords: Keyword[];
  topNegativeKeywords: Keyword[];
  topNeutralKeywords: Keyword[];
}

export interface AISummary {
  videoId: number;
  summary: string;
  strengths: string[];
  weaknesses: string[];
  commonThemes: string[];
  generatedAt: string;
}

export type QueryCommentsParams = {
sentiment?: QueryCommentsSentiment;
search?: string;
page?: number;
limit?: number;
};

export type QueryCommentsSentiment = typeof QueryCommentsSentiment[keyof typeof QueryCommentsSentiment];


export const QueryCommentsSentiment = {
  positive: 'positive',
  negative: 'negative',
  neutral: 'neutral',
  all: 'all',
} as const;