# lib/api-zod/src/generated/types/aISummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YouTube Audience Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AISummary {
  videoId: number;
  summary: string;
  strengths: string[];
  weaknesses: string[];
  commonThemes: string[];
  generatedAt: string;
}