# lib/api-zod/src/generated/types/analytics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YouTube Audience Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Keyword } from './keyword';
import type { SentimentSummary } from './sentimentSummary';
import type { SentimentTimePoint } from './sentimentTimePoint';

export interface Analytics {
  sentimentSummary: SentimentSummary;
  sentimentOverTime: SentimentTimePoint[];
  topPositiveKeywords: Keyword[];
  topNegativeKeywords: Keyword[];
  topNeutralKeywords: Keyword[];
}