# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YouTube Audience Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CommentSentiment } from './commentSentiment';

export interface Comment {
  id: number;
  videoId: number;
  text: string;
  authorName: string;
  likeCount: number;
  publishedAt: string;
  sentiment: CommentSentiment;
  confidence: number;
}