# lib/api-zod/src/generated/types/commentPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YouTube Audience Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Comment } from './comment';

export interface CommentPage {
  data: Comment[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}