# lib/api-zod/src/generated/types/keyword.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YouTube Audience Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Keyword {
  word: string;
  count: number;
}