# lib/api-zod/src/generated/types/sentimentTimePoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YouTube Audience Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SentimentTimePoint {
  date: string;
  positive: number;
  negative: number;
  neutral: number;
}