# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YouTube Audience Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoStatus } from './videoStatus';

export interface Video {
  id: number;
  youtubeId: string;
  title: string;
  channelName: string;
  viewCount: number;
  likeCount: number;
  publishedAt: string;
  thumbnailUrl: string;
  commentCount: number;
  status: VideoStatus;
  createdAt: string;
}