# lib/api-zod/src/generated/types/videoDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YouTube Audience Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SentimentSummary } from './sentimentSummary';
import type { VideoDetailStatus } from './videoDetailStatus';

export interface VideoDetail {
  id: number;
  youtubeId: string;
  title: string;
  channelName: string;
  viewCount: number;
  likeCount: number;
  publishedAt: string;
  thumbnailUrl: string;
  commentCount: number;
  status: VideoDetailStatus;
  createdAt: string;
  sentimentSummary: SentimentSummary;
}